pyproject.toml
[project]
name = "taskassistant"
version = "0.1.0"
description = "A simple REST API for managing tasks, built with FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "uvicorn>=0.34.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "google-genai>=1.0.0",
    "openai>=1.50.0",
]
